ITDS-24 refactor: 도메인 entity, enum 패키지 분리
