+++ src/main/java/com/dissonance/itit/domain/entity/BaseTime.java
@@ -1,4 +1,4 @@
-package com.dissonance.itit.domain;
+package com.dissonance.itit.domain.entity;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
