+++ src/main/java/com/dissonance/itit/domain/entity/User.java
@@ -1,13 +1,14 @@
-package com.dissonance.itit.domain;
+package com.dissonance.itit.domain.entity;
 
+import com.dissonance.itit.domain.enums.Role;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
+@Getter
 @Builder
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Table(name = "user")
